+++ packages/renderer/src/dialogs/gist-share-dialog.ts
@@ -7,8 +7,7 @@
 
 import { LitElement, css, html } from 'lit';
 import { customElement, state } from 'lit/decorators.js';
-import { ctrlEnterSubmits, dialogChromeStyles } from './dialog-chrome.js';
-import { makeDialogDraggable } from './draggable.js';
+import { ctrlEnterSubmits, dialogChromeStyles, makeDialogDraggable } from '@cawoodm/lit-dialogs';
 
 let singleton: GistShareDialog | null = null;
 function host(): GistShareDialog {